requirements.txt
Django==2.1.5
django-filter==2.2.0
djangorestframework==3.10.2
Markdown==3.1.1
Pillow==5.2.0
pytz==2018.9
